% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead0.R
\name{lead0}
\alias{lead0}
\title{Add Leading 0s to A Number}
\usage{
lead0(vect, len)
}
\arguments{
\item{vect}{Either a numeric value or a vector of numeric values to which you want to add leading 0s
len The length you want each value to have after leading 0s are added.  For example, if every number in a vector should be 10 characters long, the function will add 1 leading 0 to 123456789, but 5 leading 0s to 12345.}
}
\description{
This function will add leading 0s to a number to the values in a vector of numbers.  The result will be a character string or vector of character strings.
}
\examples{
lead0(12345, len = 10)
lead0(ids, len = 20)
lead0(sales$product_id, len = 15)
}
\keyword{sqlWriter}
