% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlIn.R
\name{sqlIn}
\alias{sqlIn}
\title{Create a List for SQL IN Clauses}
\usage{
sqlIn(vect, type = "char")
}
\arguments{
\item{vect}{A vector you want to use in an SQL IN Clause
type Either "num" for numeric or "char" for character.  "char" will put single quotes around each value in the list so the query will be process the values as strings.  Defaults to "char".}
}
\description{
This function allows you to take a vector and turn it into a character string containing a comma separated list, bookended by parentheses, for use in an SQL IN clause.
}
\examples{
sqlIn(mtcars$mpg, type = "num")
sqlIn(iris$Species)
}
\keyword{sqlWriter}
