% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlInsert.R
\name{sqlInsert}
\alias{sqlInsert}
\title{Insert Values into an SQL Table}
\usage{
sqlInsert(sourceframe, table)
}
\arguments{
\item{sourceframe}{A data frame you want to insert into a table
table The name of the table you want to insert values into.}
}
\description{
This function converts a Data Frame or Matrix into a character string containing an SQL statement to insert the values into a table.
}
\examples{
sqlInsert(mtcars, table = "car_data")
}
\keyword{sqlWriter}
