% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlLike.R
\name{sqlLike}
\alias{sqlLike}
\title{Create a Series of SQL LIKE Statements}
\usage{
sqlLike(vect, field, wildcard = "\%", loc = "end")
}
\arguments{
\item{vect}{A vector you want to use for creating your LIKE statements
field The name of the field in your SQL database to use in your LIKE statements
wildcard Which wildcard character you would like to use in your LIKE statements. Defaults to "%"
loc Where you would like to locate the wildcard character for each value.  Current options are "end" for the appending the wildcard to the end of the value, "lead" for appending the wildcard to the beginning of the value, and "both" for both.  Defaults to "end".}
}
\description{
This function allows you to take a vector and turn it into a character string containing a series of LIKE statements for use in an SQL query.
}
\examples{
sqlLike(iris$Species, field = "iris_species")
sqlLike(iris$Species, field = "iris_species", wildcard = "_", loc = "both")
}
\keyword{sqlWriter}
