% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTempTable.R
\name{sqlTempTable}
\alias{sqlTempTable}
\title{Insert Values into a Temporary SQL Table}
\usage{
sqlTempTable(sourceframe, table)
}
\arguments{
\item{sourceframe}{A data frame you want to insert into a temporary table
table The name of the temporary table you want to insert values into.}
}
\description{
This function converts a Data Frame or Matrix into a character string containing an SQL statement to insert the values into a temporary table.  While this will create all the code for inserting the values into the temporary table, you will still need to write code to create the temporary table in your query.
}
\examples{
sqlTempTable(mtcars, table = "#mtcars")
}
\keyword{sqlWriter}
