% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlUpdate.R
\name{sqlUpdate}
\alias{sqlUpdate}
\title{Update sqlWriteR}
\usage{
sqlUpdate()
}
\description{
This function simplifies updating the sqlWriteR library by re-installing it from github.
Simply run sqlUpdate() and R will update and reload sqlWriteR for you!
}
\examples{
sqlUpdate()
}
\keyword{sqlWriteR}
